requirements.txt
fastapi
uvicorn[standard]
pillow

# Vision + embeddings
torch
torchvision
sentence-transformers

# Optional but recommended for robustness
numpy

